pyproject.toml
[tool.poetry]
name = "de_ds_ai_automation"
version = "0.1.0"
description = "Data Standardization project"
authors = ["Irshad"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "1.34.84"
botocore = ">=1.34.84,<1.35.0"
pandas = "2.1.4"
pyarrow = "14.0.0"
redshift-connector = "2.0.918"
sagemaker = "2.212.0"
smart-open = "5.2.1"
pyodbc = "5.0.1"
pipeline = "^0.1.0"
polars = "0.20.6"
requests = "2.31.0"
responses = "0.18.0"
s3transfer = "0.10.1"
jupyter = "^1.0.0"
notebook = "^7.2.1"
debugpy = ">=1.8.0,<1.9.0"
azure-devops = "^7.1.0b4"
schedule = "^1.2.2"
selenium = "^4.25.0"
pre-commit = "^4.0.1"
openai = "^1.58.1"
pywhatkit = "^5.4"
pynput = "^1.8.0"
pyspark = "3.3.2"
python-docx = "^1.1.2"
markdownify = "^1.1.0"
python-dotenv = "^1.0.0"
beautifulsoup4 = "^4.12.2"
pytz = "^2024.1"
cohere = "^4.40"
fake-useragent = "^1.5.1"
nltk = "^3.9.1"
docx2pdf = "^0.1.8"
pytest = "^8.3.4"
pyautogui = "^0.9.54"
pyscreeze = "^1.0.1"
pillow = "^11.3.0"
opencv-python = "^4.11.0.86"
pytesseract = "0.3.10"
pypandoc = "^1.15"
pymupdf = "^1.26.1"
scikit-learn = "^1.7.0"
sentence-transformers = "^5.0.0"
rouge-score = "^0.1.2"
pydrive = "^1.3.1"
streamlit = "^1.47.1"
langchain = "^0.3.27"
dotenv = "^0.9.9"
unstructured = "^0.18.11"
pdfminer = "^20191125"
pdfminer-six = "20221105"
pi-heif = "^1.1.0"
unstructured-inference = "^1.0.5"
pdf2image = "^1.17.0"
chromadb = "^1.0.17"
chromadb-client = "1.0.16"
gradio = "^5.44.0"
diffusers = "^0.35.1"
safetensors = "^0.6.2"
moviepy = "^2.2.1"
pyttsx3 = "^2.99"
tqdm = "^4.67.1"
srt = "^3.5.3"
onnxruntime = "1.20.0"
whisper = "^1.1.10"
accelerate = "^1.10.1"
imageio = "^2.37.0"
imageio-ffmpeg = "^0.6.0"
soundfile = "^0.13.1"
elevenlabs = "^2.13.0"
transformers = ">=4.44"
torch = "2.4.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
